Make deleting new lines work as expected.
Since an empty space character is being put in in addition to a new line
character, there are two characters to delete whenever deleting a new
lien. This fix now checks to see if a new line is being deleted and does
an additional delete. So a single delete now deletes a new line as the
user would expect.